+++ assets/SunIcon.tsx
@@ -1,5 +1,5 @@
 //? Import IconProperties for props typecasting
-import { IconProperties } from "../types/IconProperties.ts";
+import { IconProperties } from "../types/misc/IconProperties.ts";
 
 //? Renders a sun Icon with fill
 export function SunIcon(
