+++ components/blog/BlogNavigationButtons.tsx
@@ -1,7 +1,7 @@
 //? Import LinkProperties to type what is required for the
 //? optional "back" and "next" buttons
 import { ChevronIcon } from "../../assets/ChevronIcon.tsx";
-import { LinkProperties } from "../../types/LinkProperties.ts";
+import { LinkProperties } from "../../types/misc/LinkProperties.ts";
 //? Define optional properties for the buttons
 interface BlogNavigationButtonProperties {
   back?: LinkProperties;
