+++ components/expenses-tracker/ExpensesChart.tsx
@@ -1,7 +1,7 @@
 //? Import individual MonthBar components to render all 12 months
 import { ExpensesChartMonthBar } from "./ExpensesChartMonthBar.tsx";
 //? Import the Expense type for casting
-import type { Expense } from "../../types/Expense.ts";
+import type { Expense } from "../../types/expenses-tracker/Expense.ts";
 
 //? Define properties for this component
 interface ExpenseChartProperties {
