+++ types/misc/IconProperties.ts
