+++ types/misc/LinkProperties.ts
